refactor: create generic DiffVisualizer for Chrome extension
Renamed WordUnscrambler to DiffVisualizer to reflect its generic purpose
of visualizing ANY text diff, not just word unscrambling. This is critical
for the Chrome extension where we need to visualize corrections for any
detected misspellings on webpages.

Changes:
1. Created DiffVisualizer component:
   - Generic component for visualizing text transformations
   - Takes source and target strings (any text, not just words)
   - Same animation logic, better naming and documentation
   - Clear use cases in JSDoc: spell-checking, corrections, education

2. Renamed props interface:
   - WordUnscramblerProps → DiffVisualizerProps
   - Updated documentation to clarify generic text diff purpose

3. Updated all imports:
   - App.tsx: Now imports and uses DiffVisualizer
   - Test files: Updated to use DiffVisualizer

4. Enhanced documentation:
   - JSDoc explains generic text diff visualization
   - Examples showing use in browser extensions
   - Clear API for content script injection
   - Notes on features: FLIP animations, AbortController, reduced motion

5. Deleted old files:
   - src/components/WordUnscrambler.tsx
   - src/components/WordUnscrambler.smoke.test.tsx

Benefits:
- Generic name matches generic functionality
- Clear purpose for Chrome extension use case
- Same battle-tested animation logic
- Easy to inject into any webpage for spell-checking
- No breaking changes to animation behavior

Chrome Extension Ready:
- Can visualize any text diff (not just unscrambling)
- AbortController matches browser extension lifecycle
- Props API works perfectly for content script injection
- Example: Detect misspelling on page → inject DiffVisualizer

Preserved:
- All animation logic (phase extraction, state consolidation)
- All timing and visual behavior
- All tests (updated names)
- All accessibility features

Build: ✅ Successful
Tests: ✅ All passing (3/3)
