+++ src/components/DiffVisualizer.smoke.test.tsx
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest';
 import { render, waitFor } from '@testing-library/react';
-import WordUnscrambler from './WordUnscrambler';
+import DiffVisualizer from './DiffVisualizer';
 
 // Helper to read current letters from DOM
 function textFromLetters(container: HTMLElement): string {
@@ -8,16 +8,16 @@ function textFromLetters(container: HTMLElement): string {
   return spans.map((s) => s.textContent ?? '').join('');
 }
 
-describe('WordUnscrambler smoke', () => {
+describe('DiffVisualizer smoke', () => {
   it('reaches final target text after animation', async () => {
     const source = 'tesd';
     const target = 'tads';
     const { container, rerender: rr, unmount } = render(
-      <WordUnscrambler source={source} target={target} animateSignal={0} />
+      <DiffVisualizer source={source} target={target} animateSignal={0} />
     );
 
     // Trigger animation by bumping animateSignal
-    rr(<WordUnscrambler source={source} target={target} animateSignal={1} />);
+    rr(<DiffVisualizer source={source} target={target} animateSignal={1} />);
 
     await waitFor(
       () => {
